Suppress InvocationTargetException in ActionTaker
The goal of model checking is to ensure that the invariants are honoured
through the possible interleaving of executions. The application level
exception should not be taken into consideration because they belong to
application level implementation, which should have their own testing
in terms of protecting the application from getting corrupted.