Use deep copy when merging global state
GlobalState is used to restore actors, via each of the corresponding
LocalState. When a new GlobalState is produced, if some of its
LocalStates are shared with the old one, when they are updated by an
actor's action, the change would be propagated to the other GlobalState,
corrupting the state space.

This commit changes how new GlobalState is produced when merging
happens. It now makes a deep copy of the existing state so no more data
is shared between states.