Make GlobalState immutable
GlobalState is used to restore actors. When restoring, the GlobalState
becomes part of the actor and the state itself changes as the actor
takes action (proactively or passively). If the GlobalState is mutated
due to sharing of data, the mutation is then propagated to other actors
which may result in a corruption of the state.

The correct way of doing this is by not sharing any data when
GlobalState is being used to restore actors. This commit changes the
interface of GlobalState so access to the internal states results in a
deep copy of the state.

This change however, mixes data and behaviour, which is not ideal. We
shall review this later:
https://github.com/stephen-hlx/overcooked/issues/6