Redesign TransactionManager and ResourceManager
The existing design of both TransactionManager and ResourceManager is
not really creating an actor that represents both the client and server.
However, in order to have a single actor in the state machine that is
able to perform proactive action and also be an action receiver, they
need to be client and server at the same time. This commit redesigns
them.