Use type info from ParamTemplate to materialise it
The ParamTemplate comes with type information in it. However, the
current implementation of the ActionTemplateMaterialiser does not use it
to convert it to a concrete ParamValue, which is incorrect as the
provided value may be a derived class of the designated type. The
consequence is that when the ActionTaker tries to find the method to
invoke, the type information retrieved from the action receiver will not
match the one defined in the action taker's interface.

For instance:
ParamTemplate has type `Base`
The action definition could be something `method(Base base)`

But when it is invoked, it could be using `method(derived)` where
`derived` is of type `Derived implements Base`.