Rename StateMachineExecutorDataCollector to ExecutionContext
This is one of the first few steps to redesign how the
StateMachineDriver is going to work using the context data. The reason
behind it is that whenever a new GlobalState is created (via merging),
it is necessary to find out if it is a GlobalState that has been seen
before and therefore the transition recorded should be using the
existing GlobalState instead of the new one.

In other words, a pool of GlobalStates needs to be maintained in order
to avoid having duplicate GlobalStates.