Replace GlobalState pool's HashMap impl using HashSet
GlobalState pool's HashMap based implementation is flawed because it is
hard to guarantee a collision free hash implemenation for GlobalState,
and this leads to overwriting GlobalState in the pool, and that means a
loss of GlobalState.

This commit replaces the HashMap using HashSet. Although less efficient
than using HashMap, it is correct.