Decouple sample from ActionTakerTest
Sample data is for model checking. It should and will be moved into its
own module. The use of sample data in core testing creates a dependency
on the sample data which should be indenpendent from each other.

This commit replaces the sample used in ActionTakerTest with dedicated
test data.