Add ActionResult for ActionTaker
In order to capture possible execution failures, the interface of
ActionTaker is updated to return the result, which could include the
exception if there is any. This information will be captured in a new
structure called ActionResult.

This information will later be used in the verbosity control of listing
the transitions for building the graph. This can help reduce the noise
in the graph so it will be eaiser for analysis.