Change parameter in ActionTemplate from list to a single value
The ActionTemplate used to have a list of parameters in it. The decision
was incorrectly driven by that the Class.getMethod requires a list of
parameter types. However, the use case for the project never needed more
than one parameter. This commit changes the type from list to a single
value.

This commit also touches up a little of the tests to remove the
dependency on the sample package.