Use lambda to replace the use of reflection
It turns out the implementation of method calling can be simplified by
replacing the use of reflection with lambda. That way, the type
information of the callee becomes moot and therefore the existing
`Param` concept can be removed entirely.

This also helps strengthen the types by reducing the use of `Object`.
Although it does not eradicate `Object` completely, it is a step
forward.