Use generic to replace the use of Object type
Object as a type has been widely used in the implementation to get
around some flaw in the typing system in this library. This commit
replaces them with the use of generics. The change here although is not
perfect as it inevitably has to resort to the use of
`SuppressWarning("unchecked")`, which is also used previously, it does
strengthen the typing in the implementation of the library.