Rename methodName to actionLabel
The term "methodName" implies a formal ID of a method to be invoked.
However, since the switch from reflection to lambda, the name here has
become more of an FYI thing. This commit changes that to "label", which
is more clear that it only really is a label, and does not imply
anything that affects how the system works.