Rename Actor to ActorId
The class Actor was initially named ActorDefinition, containing the type
information of the actor object. After a few iterations of design
changes, this is no longer needed and it got renamed to Actor. Further
refactoring has its responsibility stripped down to leaving only the ID
of an actor. Although it has only a String attribute, it is strongly
typed as a class. This commit changes its name from Actor to ActorId.