Refactor ResourceManager and TransactionManger
Both ResourceManager and TransactionManager are a composition of the
client and server implementation of the ResourceManager and
TransactionManager. But the implementations for model verification were
carelessly named as a simple implementation instead of what they
actually are, an in-memory implementation. This commit fixes that by
renaming them.

This commit also introduces a `RefCell` concept borrowed from rust so
that the same copy of state can be easily shared across different owners
of the state. The in-memory implementations as well as the
ResourceManager and TransactionManager are now using this new structure
to share the state with its internal / composite structure, the client
and server in-memory implementations.