Use sample package for testing lib
The lib package currently does not have its own end-to-end testing. This
commit makes use Jar sample in the sample package to do an end-to-end
test for the lib package.

It does create a little duplicate code between the lib's test code and
the sample package, but it makes the lib package more self-contained in
terms of testing.