Rename Two Phase Commit sample classes
This commit renames following classes:
- ResourceManager -> ResourceManagerActor
- ResourceManagerServer -> ResourceManager
- ResourceManagerLocalStateExractor ->
  ResourceManagerActorStateExtractor
- TransactionManager -> TransactionManagerActor
- TransactionManagerServer -> TransactionManager
- TransactionManagerLocalStateExtractor ->
  TransactionManagerActorLocalStateExtractor

The reason of this renaming is that, the `%Server` on its own is not
appropriate as it implies that the instance should include the server
stack, e.g. communication protocol, which is not the concern of model
verification. Removing the `Server` from their names makes more sense as
the names now focus more on its behaviour instead of the whole server
stack.

On the other hand, this commit also appends `Actor` to the names of the
classes that are used for the actors in the model verification.