Refactor sample ResourceManager interface
This commit renames `ResourceManager.abort()` to
`ResourceManager.selfAbort()``.

Both `ResourceManager` and `ResourceManagerClient` have the interface of
`abort()`:
- For `ResourceManager`, it is to initiate the abort by itself
- For `ResourceManagerClient`, it is to let the `TransactionManager`
  abort the transaction

Having the same name for both of them makes it a bit confusing in
discussions as additional comments will be needed to tell them apart,
like proactive and passive. By renaming it, it removes the ambiguity.`