Add ModelVerifier
The `StateMachine` is an implementation details of the model
verification. However, it is currently exposed by the library as an API
which is totall unnecessary.

This commit adds the `ModelVerifier` which encapsulates these
implementation details.

This commit also refactors the sample model verification by moving them
to the unit test.