Rename sample diehard to waterjar
The sample diehard was inspired by TLA+'s tutorial examples. However,
without context, it is not a very straightforward for whoever does not
have the context about the name of this example. Instead, waterjar is a
pretty well known puzzle and using waterjar for the sample name makes
things simpler.

This commit simply renames the package to waterjar.