Make `equals` a mandatory method for `LocalState` implementations
Currently, it is not obvious that the `equals` method is mandatory for
`LocalState` implementations. This may lead to unexpected results when 2
`LocalState` implementations are compared against each other.

This commit adds the `equals` method to the abstract class of
`LocalState`, mandating all its implementations to provide an
implementation of it.