Rename LocalState to ActorState
The reason it started with the name `LocalState` was being relative to
`GlobalState`. However, it really is only about the actor. So
`ActorState` might have been more accurately.

Moreover, there will be some upcoming change that is about adding
metadata to an actor that is about its state in the distributed system,
which is outside of the actor's internal state, e.g. whether it has died
/ disconnected from the (partially) the group.

This commit is a pre-requisite for the upcoming change.