Add LocalState and ActorEnvState
The current ActorState represents the internal state of an actor. But
this is not sufficient to represent state like the reachability of the
actor, which is determined by some property outside of the actor, e.g.
network, state of the container, etc.

This commit adds ActorEnvState, which would encapsulate the
aforementioned environment related state of an actor. And along with
ActorState, the internal state, they will be contained in the
LocalState, which holds all information locally to the environment
where the actor lives.