Rename the sampe WaterJar to WaterJug
I just realised that Jar actually means something in Java. It would be
confusing for people looking into this library for the first time as by
"Jar" it could mean the "Java Archive" and that might mislead people to
think that this library could directly test a Jar file, which is not the
case here.

This commit, although creates quite some noise in the code history,
renames the sample name from "WaterJar" to "WaterJug". All code
references to "Jar" in any cases have been updated.
"