Refactor the TwoPhaseCommit sample
This commit refactors the server side implementation of both
`TransactionManager` and `ResourceManager`.

The imaginary prod version of the implementation likely has a
non-volatile storage of the state of the application. The so-called
in-memory implementation is a simplified version that makes the
restoration of the application easier. Therefore, the part of the
implementation required to be in-memory is just the storage of the
state.

This commit further decouples this in-memory concept from the
implementation of the `ResourceManager` by having a
`ResourceManagerStateDao`. As for `TransactionManager`, it simply uses a
map to store the state of the `ResourceManager`s in memory, for the sake
of simplicity, no DAO is made for it.